build: ensure only one package's coverage is evaluated at a time
We were encountering a bug due to the HTML code coverage reports flattening their directory structure when all dependencies of the tested package reside in the same namespace
as the package itself.
While we could preserve the full project package structure with the  --all option, setting it would be too expensive. Hence, we instead only generate the test coverage for
the currently tested package by setting the appropriate c8 flags (for which the make command recipe had to be updated to accept custom flags) such that we can sidestep the
need to resolve the correct folder to upload to the code coverage repository, Our previous approach of finding the folder for the tested package was not satisfactory as it
didn't take the flattening behavior into account and the index.html file existence check was thus not always yielding the corect result.

---
type: pre_commit_static_analysis_report
description: Results of running static analysis checks when committing changes.
report:
  - task: lint_filenames
    status: passed
  - task: lint_editorconfig
    status: passed
  - task: lint_markdown
    status: na
  - task: lint_package_json
    status: na
  - task: lint_repl_help
    status: na
  - task: lint_javascript_src
    status: na
  - task: lint_javascript_cli
    status: na
  - task: lint_javascript_examples
    status: na
  - task: lint_javascript_tests
    status: na
  - task: lint_javascript_benchmarks
    status: na
  - task: lint_python
    status: na
  - task: lint_r
    status: na
  - task: lint_c_src
    status: na
  - task: lint_c_examples
    status: na
  - task: lint_c_benchmarks
    status: na
  - task: lint_c_tests_fixtures
    status: na
  - task: lint_shell
    status: passed
  - task: lint_typescript_declarations
    status: passed
  - task: lint_typescript_tests
    status: na
  - task: lint_license_headers
    status: passed
---