refactor!: rename `sentinelValue` option in `ndarray/find`
This commit renames the `sentinelValue` option to `sentinel`, with the latter being more succinct and arguably more ergonomic. The use of `Value` did not add anything to the name, as value is already implied by the use of `sentinel`.

BREAKING CHANGE: rename `sentinelValue` option to `sentinel`

To migrate, users should update their usage of `sentinelValue` to use just `sentinel`.